Fix unknown label
